+++ web/src/main.ts
@@ -1,16 +1,15 @@
 import { createPinia } from "pinia";
-import { makeRouter } from "./router";
+import { createRouter } from "./router";
 import { library } from "@fortawesome/fontawesome-svg-core";
 import { faUserSecret } from "@fortawesome/free-solid-svg-icons";
-import { mount } from "./mount";
-import { configs as appConfigs } from "./configs";
-import { configs as routerConfigs } from "./router/configs";
+import { createApp } from "./mount";
 import "@/styles/main.scss";
 
 import type {
   Library,
   IconDefinition,
 } from "@fortawesome/fontawesome-svg-core";
+import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 
 function setupFontAwesomeLibrary(
   library: Library,
@@ -21,13 +20,15 @@ function setupFontAwesomeLibrary(
 
 setupFontAwesomeLibrary(library, [faUserSecret]);
 
-mount(
-  {
-    router: makeRouter(),
-    pinia: createPinia(),
-  },
-  {
-    appConfigs: appConfigs,
-    routerConfigs: routerConfigs,
-  }
-);
+function mount(): void {
+  const app = createApp();
+  const router = createRouter();
+  const pinia = createPinia();
+
+  app.use(router);
+  app.use(pinia);
+  app.component("FontAwesomeIcon", FontAwesomeIcon);
+  app.mount("#app");
+}
+
+mount();
