+++ web/src/router/index.ts
@@ -1,9 +1,10 @@
-import { createRouter, createWebHistory } from "vue-router";
+import { createRouter as vueRouterCreate, createWebHistory } from "vue-router";
 
-import type { Router } from "vue-router";
+import type { NavigationGuard, Router } from "vue-router";
 import { routes } from "./routes";
 import { configs } from "./configs";
-import type { RouterConfigs } from "./configs";
+
+export type RouterConfigs = Record<string, NavigationGuard>;
 
 export function addConfigs(router: Router, configs: RouterConfigs): Router {
   if (configs.beforeEach) router.beforeEach(configs.beforeEach);
@@ -11,8 +12,8 @@ export function addConfigs(router: Router, configs: RouterConfigs): Router {
   return router;
 }
 
-export function makeRouter() {
-  const router = createRouter({
+export function createRouter() {
+  const router = vueRouterCreate({
     history: createWebHistory(import.meta.env.BASE_URL),
     routes: routes,
   });
